int replace

+++ include/dspatch/DspThreadNull.h
@@ -69,7 +69,7 @@ public:
     static void SetPriority(Priority priority)
     {
     }
-    static void MsSleep(unsigned short milliseconds)
+    static void MsSleep(int milliseconds)
     {
     }
 };
